Update files
I added a quote from Sam Gamgee to the README.md file.
I added an endpoint at /api/number to the main.go file to return a random number between 20 and 50.
I modified the index.html file to have a different color for the content section compared to the background.

Resolves: #5