Add a "numbers" endpoint
Added a new "/api/numbers" endpoint in main.go that returns a random integer between 14 and 37.
Updated index.html with a content section with a slightly different background color and modified the button to make a request to the new endpoint. The result is displayed in the body of the page.

Resolves: #15