Set up basic server
I have added an HTML page with an input field to submit a number and an additional button to retrieve the previously submitted number. I have also added JavaScript code to handle these events. In main.go, I have set up a server that listens on port 8080. It serves index.html on the root route and accepts POST requests to a /number endpoint to store a shared integer and GET requests to retrieve the same integer.

Resolves: #34