Make some initial files
The index.html file was modified by adding the required content, styling, and flexbox properties for centering. The body has a light blue background color (#E6F1F8) and the content section has an off-white background color (#FAFAFA) with a box-shadow and max-width of 600px. The main.go file was modified by adding an endpoint for a GET request to /api/number that returns a random number between 12 and 34.

Resolves: #36