initialize simple oauth app
I have created the main.go file to serve index.html and support OAuth2 authentication via GitHub. When the user is authenticated, the access token will be logged. The index.html file is a simple landing page for a product with specified styles. I have also added a Dockerfile to run the server from a container and included instructions to run the program with `go` and `docker` in the README.md.

Resolves: #38