Write a home page and simple http server
In this task, I've created an html page [index.html] which includes a button to fetch a random number from the server. On the other hand a Go server [main.go] has been created to serve this static file and an API endpoint /get-number to return a random number between 23 and 50. When number gets fetched from server, it is inserted into html page.
For alignment adjusting and responsiveness, CSS Flexbox is being used. The background colors are soft shades of grey. Hope this fits the requirement well.

Resolves #42